cell-item-impl

+++ app/src/main/java/com/ashana/test/coolblue/lite/network/model/ProductDTO.kt
@@ -21,6 +21,6 @@ class ProductDTO(
     @SerializedName("reviewCount")
     var reviewCount: Int? = null,
     @SerializedName("reviewAverage")
-    var reviewAverage: Float? = null
+    var reviewAverage: Double? = null
 
 )
